backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.2.0
gunicorn>=21.2.0
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
